♻️ refactor: 토큰 쿠키 관련 함수들 모델로 이동
